https://codeobia.atlassian.net/browse/PLA-211

+++ back-end/src/melia-type/melia-type.controller.ts
@@ -1,6 +1,7 @@
-import { Body, Controller, Delete, Get, Param, Post, Put, Query } from '@nestjs/common';
+import { BadRequestException, Body, Controller, Delete, Get, Param, Post, Put, Query } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { MeliaTypes } from 'src/entities/melia-types.entity';
+import { Melia } from 'src/entities/melia.entity';
 import { ILike, Repository } from 'typeorm';
 
 @Controller('melia-type')
@@ -8,6 +9,8 @@ export class MeliaTypeController {
     constructor(
         @InjectRepository(MeliaTypes)
         private meliaTypesRepository: Repository<MeliaTypes>,
+        @InjectRepository(Melia)
+        private MeliaRepository: Repository<Melia>,
       ) {}
 
     @Get('')
@@ -25,7 +28,14 @@ export class MeliaTypeController {
       return this.meliaTypesRepository.findOne({where: { id: id }});
     }
     @Delete(':id')
-    remove(@Param('id') id: string) {
+    async remove(@Param('id') id: number) {
+      const melia = await this.MeliaRepository.find({
+        where: {
+          melia_type: id
+        }
+      });
+      if(melia.length)
+        throw new BadRequestException('MELIA type can not be deleted, This MELIA type is assigned for an initiative(s)')
       return this.meliaTypesRepository.delete(id);
     }
